consolidate helpers
